+++ source/painters/stdnotationnote.cpp
@@ -85,8 +85,7 @@ void StdNotationNote::getNotesInStaff(
     const Staff &staff, int staffIndex, const LayoutInfo &layout,
     std::vector<StdNotationNote> &notes,
     std::array<std::vector<NoteStem>, Staff::NUM_VOICES> &stemsByVoice,
-    std::array<std::vector<BeamGroup>, Staff::NUM_VOICES> &groupsByVoice,
-    const QColor &notesColor)
+    std::array<std::vector<BeamGroup>, Staff::NUM_VOICES> &groupsByVoice)
 {
     // If there is no active player, use standard 8-string tuning as a default
     // for calculating the music notation.
@@ -216,7 +215,7 @@ void StdNotationNote::getNotesInStaff(
                     NoteStem(voice, pos, x, noteHeadWidth, noteLocations));
             }
 
-            computeBeaming(bar.getTimeSignature(), stems, firstStem, groups, notesColor);
+            computeBeaming(bar.getTimeSignature(), stems, firstStem, groups);
         }
 
         voiceIndex++;
@@ -311,8 +310,7 @@ std::vector<uint8_t> StdNotationNote::getBeamingPatterns(
 void StdNotationNote::computeBeaming(const TimeSignature &timeSig,
                                      std::vector<NoteStem> &stems,
                                      size_t firstStemIndex,
-                                     std::vector<BeamGroup> &groups,
-                                     const QColor &beamingColor)
+                                     std::vector<BeamGroup> &groups)
 {
     const std::vector<uint8_t> beamingPatterns(getBeamingPatterns(timeSig));
 
@@ -349,7 +347,7 @@ void StdNotationNote::computeBeaming(const TimeSignature &timeSig,
         computeBeamingGroups(stems, durations, subgroupLength, firstStemIndex,
                              firstStemIndex + (groupStart - durations.begin()),
                              firstStemIndex + (groupEnd - durations.begin()),
-                             groups, beamingColor);
+                             groups);
 
         // Move on to the next beaming pattern, looping around if necessary.
         ++groupSize;
@@ -363,8 +361,7 @@ void StdNotationNote::computeBeaming(const TimeSignature &timeSig,
 void StdNotationNote::computeBeamingGroups(
     std::vector<NoteStem> &stems, const std::vector<double> &durations,
     const std::optional<double> &subgroupLength, size_t firstStemIndexInBar,
-    size_t firstStemIndex, size_t lastStemIndex, std::vector<BeamGroup> &groups,
-    const QColor &beamColor)
+    size_t firstStemIndex, size_t lastStemIndex, std::vector<BeamGroup> &groups)
 {
     // Rests and notes greater than eighth notes will break apart a beam group,
     // so we need to find all of the subgroups of consecutive positions that
@@ -381,7 +378,7 @@ void StdNotationNote::computeBeamingGroups(
             if (NoteStem::needsStem(stems[i]))
             {
                 groups.push_back(
-                    BeamGroup(NoteStem::formatGroup(stems, { i }), { i }, beamColor));
+                    BeamGroup(NoteStem::formatGroup(stems, { i }), { i }));
             }
 
             ++i;
@@ -396,7 +393,7 @@ void StdNotationNote::computeBeamingGroups(
             if (stems[i].isGraceNote())
             {
                 groups.push_back(
-                    BeamGroup(NoteStem::formatGroup(stems, { i }), { i }, beamColor));
+                    BeamGroup(NoteStem::formatGroup(stems, { i }), { i }));
             }
             else
             {
@@ -428,7 +425,7 @@ void StdNotationNote::computeBeamingGroups(
         if (!group_stems.empty())
         {
             auto direction = NoteStem::formatGroup(stems, group_stems);
-            groups.push_back(BeamGroup(direction, group_stems, beamColor));
+            groups.push_back(BeamGroup(direction, group_stems));
         }
     }
 }