+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Thanks to the following contributors who worked on this release:
 - Added support for volume swells (#17).
 - Added MIDI playback support for alterations of pace (#18).
 - Added new commands (`Note -> Shift String Up` and `Note -> Shift String Down`) to shift notes up or down by a string (#20).
+- Ties, hammer ons, pull offs, and slides can now be created between notes in adjacent systems (#27).
 - Added a dark theme for the score view, based on the user's system theme (#286, @mostafa-nabil).
 - Left hand fingerings now support thumbs. This is also now supported by the Guitar Pro 6/7 importers (#292).
 - The song information can now store a subtitle. This is also now supported by the Guitar Pro 6/7 importers (#297).